refactor: move ReadAgentTypeFromTree tests into common_test.go
No need for a standalone test file — these tests belong with the
function they test in common.go.

Entire-Checkpoint: 751884f8707a
